Docs: Enhance error messages, README, and expressions\n\n- Updated `src/errors/mod.rs` to include detailed documentation on the error philosophy and variants.\n- Revamped `README.md` with a "Hero's Journey" tutorial, Rosetta Stone, and Troubleshooting guide.\n- Added doctests to `src/semantic/expressions.rs` for `analyze_argument_expr` and documented the recursion limit.\n- Moved a regression test from `tests/` to `src/` to fix a visibility issue.\n- Updated `Bard's Journal` with a new entry on integration tests.
